requirements.txt
google-generativeai
sentence-transformers
faiss-cpu
dotenv
numpy 